requirements.txt
# Core dependencies
PyGithub>=2.1.1
openai>=1.6.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
requests>=2.31.0

# Utilities
click>=8.1.7
rich>=13.7.0
